+++ cinebench_2024/cinebench_2024.toml
@@ -1,5 +1,5 @@
 friendly_name = "Cinebench 2024"
-py_script = "cinebench.py"
+py_script = "cinebench_2024.py"
 executable = "Cinebench.exe"
 presentmon_enabled = false
 non_game = true
